+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- On Android, implement `Window::request_redraw`
 - **Breaking:** On Web, remove the `stdweb` backend.
 - Added `Window::focus_window`to bring the window to the front and set input focus.